requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
dnspython==2.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
idna==3.2
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
Pillow==8.3.1
psutil==5.8.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-Levenshtein==0.12.2
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-file==1.5.1
sgmllib3k==1.0.0
Shapely==1.7.1
six==1.16.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.2
unicode==2.8
Unidecode==1.2.0
urllib3==1.26.6
vncorenlp==1.0.3
